# Improve dictionary map performance on category series, fixes #23785