--- a/pandas/core/base.py
+++ b/pandas/core/base.py
@@ -18,7 +18,7 @@ from pandas.util._validators import validate_bool_kwarg
 from pandas.core.dtypes.common import (
     is_datetime64_ns_dtype, is_datetime64tz_dtype, is_datetimelike,
     is_extension_array_dtype, is_extension_type, is_list_like, is_object_dtype,
-    is_scalar, is_timedelta64_ns_dtype)
+    is_scalar, is_timedelta64_ns_dtype, is_categorical_dtype)
 from pandas.core.dtypes.generic import ABCDataFrame, ABCIndexClass, ABCSeries
 from pandas.core.dtypes.missing import isna
 
@@ -1187,6 +1187,11 @@ class IndexOpsMixin:
                 values = self._values
             else:
                 values = self.values
+
+            if is_categorical_dtype(values):
+                # use the built in categorical series mapper which saves
+                # time by mapping the categories instead of all values
+                return values.map(mapper)
 
             indexer = mapper.index.get_indexer(values)
             new_values = algorithms.take_1d(mapper._values, indexer)